[enhancement/fix] Begin to properly intergrate the hardened allocator to the WooKey project in the std lib. The compatibility was previously broken, and this commit is a first attempt to migrate/integrate the code.
The hardened allocator (contrary to the previously used "light" one) embeds
mitigations such as randomized allocation, canaries, heap consistency checks,
...

This is still in "beta test and review" and is a work in progress. When everything
is tested to be OK, the hardened allocator will be made as the "default" one in
the project configuration files.